Fix button
Set line-height to 1.15 explicitly, which was
set differently depending on whether the
button was rendered as a button or a elem.
